refactor: Move searching to internal/
