+++ internal/searching/binary.go
